# src/main/resources/static/openapi.yaml
openapi: 3.0.0
info:
  title: Project API
  description: API for booking dates to skate on a skatepark.
  version: 1.0.0
  contact:
    name: Luis Felipe
    email: lipalhos@gmail.com
servers:
  - url: http://localhost:8080/reservation
    description: Local server

tags:
  - name: Auth Service
    description: Endpoints for authentication and user registration
  - name: Scheduling Service
    description: Endpoints for scheduling management
  - name: Scheduling Analytics
    description: Aggregated scheduling data for dashboards
  - name: Track Service
    description: Endpoints for track management
  - name: User Service
    description: Endpoints for user management

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        cpf:
          type: string
        role:
          type: string
        encodedPassword:
          type: string

    UserRegisterDto:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        cpf:
          type: string
        password:
          type: string
        role:
          type: string

    UserEditDto:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        cpf:
          type: string
        password:
          type: string
        role:
          type: string

    UserLoginDto:
      type: object
      properties:
        email:
          type: string
        password:
          type: string

    AuthResponseDto:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
        token:
          type: string

    Payment:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string

    Turno:
      type: string
      enum: [ MATUTINO, VESPERTINO, NOTURNO ]
      example: VESPERTINO

    CheckinStatus:
      type: string
      enum: [ PENDENTE, NAO_REALIZADO, REALIZADO, CANCELADO ]
      example: PENDENTE

    CreateSchedulingDto:
      type: object
      required:
        - userId
        - trackId
        - scheduledDate
        - turno
      properties:
        id:
          type: integer
          example: 10
          description: Usado apenas para update (PUT)
        userId:
          type: integer
          example: 1
        paymentId:
          type: integer
          nullable: true
          example: null
        trackId:
          type: integer
          example: 2
        scheduledDate:
          type: string
          format: date
          example: "2025-11-10"
        turno:
          $ref: '#/components/schemas/Turno'
        checkinStatus:
          $ref: '#/components/schemas/CheckinStatus'
          description: Campo opcional usado em respostas e pode ser enviado em atualizações para alterar o status (regras de negócio se aplicam).

    Scheduling:
      type: object
      properties:
        id:
          type: integer
          example: 10
        user:
          $ref: '#/components/schemas/User'
        payment:
          $ref: '#/components/schemas/Payment'
          nullable: true
        track:
          $ref: '#/components/schemas/Track'
        scheduledDate:
          type: string
          format: date
          example: "2025-11-10"
        turno:
          $ref: '#/components/schemas/Turno'
        checkinStatus:
          $ref: '#/components/schemas/CheckinStatus'
    SchedulingTurnoSummary:
      type: object
      properties:
        turno:
          $ref: '#/components/schemas/Turno'
        total:
          type: integer
          format: int64
    SchedulingDaySummary:
      type: object
      properties:
        date:
          type: string
          format: date
        turnos:
          type: array
          items:
            $ref: '#/components/schemas/SchedulingTurnoSummary'
    SchedulingTrackSummary:
      type: object
      properties:
        trackId:
          type: integer
        trackName:
          type: string
        total:
          type: integer
          format: int64
    SchedulingPaymentSummary:
      type: object
      properties:
        paymentId:
          type: integer
          format: int64
          nullable: true
        paymentName:
          type: string
          nullable: true
        total:
          type: integer
          format: int64
    SchedulingUserSummary:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        userName:
          type: string
        total:
          type: integer
          format: int64
    SchedulingTimelinePoint:
      type: object
      properties:
        period:
          type: string
          example: "2025-01"
        total:
          type: integer
          format: int64
    SchedulingCancellationStats:
      type: object
      properties:
        months:
          type: integer
        total:
          type: integer
          format: int64
        cancelled:
          type: integer
          format: int64
        percentage:
          type: number
          format: double

    Track:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string

paths:
  /auth/register:
    post:
      tags:
        - Auth Service
      summary: Registers a new user
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegisterDto'
      responses:
        '200':
          description: User successfully registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponseDto'
        '500':
          description: Server error

  /auth/login:
    post:
      tags:
        - Auth Service
      summary: Authenticates a user
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLoginDto'
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponseDto'
        '500':
          description: Server error


  /scheduling/:
    post:
      tags:
        - Scheduling Service
      summary: Creates a new schedule (date + turno)
      operationId: createScheduling
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSchedulingDto'
            example:
              userId: 1 # ID do usuario que esta criando o agendamento
              trackId: 2 # ID da pista que esta sendo agendada 
              paymentId: 3 # ID do pagamento associado (pode ser null)
              scheduledDate: "2026-01-01" # Data do agendamento
              turno: "VESPERTINO" # OPCOES: MATUTINO, VESPERTINO, NOTURNO
              "checkinStatus": "PENDENTE" #VALOR PADRAO PARA NOVOS AGENDAMENTOS
      responses:
        '200':
          description: Schedule successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Scheduling'
        '500':
          description: Server error

    put:
      tags:
        - Scheduling Service
      summary: Updates an existing schedule
      operationId: updateScheduling
      security:
        - bearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSchedulingDto'
            example:
              id: 10
              userId: 1
              trackId: 2
              paymentId: null
              scheduledDate: "2025-11-11"
              turno: "MATUTINO"
              "checkinStatus": "PENDENTE"
      responses:
        '200':
          description: Schedule successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Scheduling'
        '500':
          description: Server error

    get:
      tags:
        - Scheduling Service
      summary: Lists all schedules
      operationId: listSchedulings
      security:
        - bearerAuth: [ ]
      responses:
        '200':
          description: Schedule list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Scheduling'
        '500':
          description: Server error

  /scheduling/{id}:
    delete:
      tags:
        - Scheduling Service
      summary: Deletes a schedule by ID
      operationId: deleteSchedulingById
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Schedule successfully deleted
        '500':
          description: Server error

  /scheduling/analytics/day:
    get:
      tags:
        - Scheduling Analytics
      summary: Retorna a quantidade de agendamentos por turno em um dia específico
      operationId: getSchedulingSummaryByDay
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: date
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Totais por turno retornados com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchedulingDaySummary'
        '500':
          description: Server error

  /payment/:
    get:
      tags:
        - Payment Service
      summary: Lists all payments
      operationId: listPayments
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Payment list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Payment'
        '500':
          description: Server error

  /scheduling/analytics/by-track:
    get:
      tags:
        - Scheduling Analytics
      summary: Retorna a quantidade total de agendamentos por pista
      operationId: getSchedulingSummaryByTrack
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: startDate
          required: false
          schema:
            type: string
            format: date
        - in: query
          name: endDate
          required: false
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Totais por pista retornados com sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SchedulingTrackSummary'
        '500':
          description: Server error

  /scheduling/analytics/by-payment:
    get:
      tags:
        - Scheduling Analytics
      summary: Retorna a quantidade total de agendamentos por forma de pagamento
      operationId: getSchedulingSummaryByPayment
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: startDate
          required: false
          schema:
            type: string
            format: date
        - in: query
          name: endDate
          required: false
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Totais por pagamento retornados com sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SchedulingPaymentSummary'
        '500':
          description: Server error

  /scheduling/analytics/by-user:
    get:
      tags:
        - Scheduling Analytics
      summary: Retorna o ranking de usuários por quantidade de agendamentos
      operationId: getSchedulingSummaryByUser
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            default: 10
        - in: query
          name: startDate
          required: false
          schema:
            type: string
            format: date
        - in: query
          name: endDate
          required: false
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Totais por usuário retornados com sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SchedulingUserSummary'
        '500':
          description: Server error

  /scheduling/analytics/timeline:
    get:
      tags:
        - Scheduling Analytics
      summary: Retorna a quantidade de agendamentos por mês em um período
      operationId: getSchedulingTimeline
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: startDate
          required: false
          schema:
            type: string
            format: date
        - in: query
          name: endDate
          required: false
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Timeline retornada com sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SchedulingTimelinePoint'
        '500':
          description: Server error

  /scheduling/analytics/cancellations:
    get:
      tags:
        - Scheduling Analytics
      summary: Retorna o percentual de cancelamentos no período informado
      operationId: getSchedulingCancellationStats
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: startDate
          required: false
          schema:
            type: string
            format: date
        - in: query
          name: endDate
          required: false
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Percentual de cancelamentos retornado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchedulingCancellationStats'
        '500':
          description: Server error

  /scheduling/analytics/day-range:
    get:
      tags:
        - Scheduling Analytics
      summary: Retorna o resumo diário por turno em um intervalo
      operationId: getSchedulingSummaryByDayRange
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: startDate
          required: false
          schema:
            type: string
            format: date
        - in: query
          name: endDate
          required: false
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Resumo diário retornado com sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SchedulingDaySummary'
        '500':
          description: Server error

  /track/:
    get:
      tags:
        - Track Service
      summary: Lists all tracks
      operationId: listTracks
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Track list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Track'
        '500':
          description: Server error

  /user/:
    put:
      tags:
        - User Service
      summary: Updates a user
      operationId: updateUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserEditDto'
      responses:
        '200':
          description: User successfully updated
        '500':
          description: Server error
    get:
      tags:
        - User Service
      summary: Lists all users
      operationId: listUsers
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '500':
          description: Server error

  /user/search:
    get:
      tags:
        - User Service
      summary: Retrieves a user by Name
      operationId: getUserByName
      security:
        - bearerAuth: [ ]
      parameters:
        - name: name
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '500':
          description: Server error


  /user/id/{id}:
    get:
      tags:
        - User Service
      summary: Retrieves a user by ID
      operationId: getUserById
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '500':
          description: Server error
    delete:
      tags:
        - User Service
      summary: Deletes a user by ID
      operationId: deleteUserById
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User successfully deleted
        '500':
          description: Server error